+++ src/TallyConnector.Models/Common/Pagination/PaginatedResponse.cs
@@ -2,7 +2,7 @@
 namespace TallyConnector.Models.Common.Pagination;
 public class PaginatedResponse<T> : PaginationBase, IPaginatedResponse<T> where T : IBaseObject
 {
-    public PaginatedResponse(int totalCount, int pageSize, List<T>? data, int pageNum) : base(totalCount, pageSize, pageNum)
+    public PaginatedResponse(ulong totalCount, int pageSize, List<T>? data, int pageNum) : base(totalCount, pageSize, pageNum)
     {
         Data = data ?? [];
     }
